Get the IPv6 scenario working.
This does a couple of things:

1) Makes vagrant scripts more reliable by attempting to create br-int.
This can prevent some errors where br-int is not there yet when
ovs-vsctl commands that use br-int are run.
2) Adds a lot of manual IPv6 stuff. This includes adding the IP
addresses to the switch ports, and adding default routes to all the
nodes. We don't have built-in IPv4 niceties, so a lot of manual
configuration is necessary here.